added click ans shape funcion
