+++ app/assets/stylesheets/application.scss
@@ -10,3 +10,6 @@
 // Your CSS partials
 @import "components/index";
 @import "pages/index";
+
+
+
