+++ app/views/layouts/application.html.erb
@@ -5,15 +5,16 @@
     <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
     <%= csrf_meta_tags %>
     <%= csp_meta_tag %>
-    <script src="libraries/p5.min.js"></script>
-  <script src="libraries/p5.sound.min.js"></script>
     <%= stylesheet_link_tag "application", "data-turbo-track": "reload" %>
+
     <%= javascript_importmap_tags %>
   </head>
 
-  <body><%= render "shared/navbar" %>
-<%= render "shared/flashes" %>
+  <body>
 
+    <%= render "shared/navbar" %>
+
+    <%= render "shared/flashes" %>
     <%= yield %>
   </body>
 </html>
