--- app/views/pages/testp5.html.erb
@@ -1,4 +1,0 @@
-<h1>test</h1>
-
-
-    <script src="/app/javascript/controllers/prueba.js"></script>
