+++ app/views/shared/_navbar.html.erb
@@ -13,7 +13,7 @@
         <% if user_signed_in? %>
 
             <li class="nav-item active">
-            <%= link_to "P5", testp5_path, class: "nav-link" %>
+            <%= link_to "P5", testpfive_path, class: "nav-link" %>
           </li>
            <li class="nav-item active">
             <%= link_to "testproject", testproject_path, class: "nav-link" %>
@@ -26,7 +26,7 @@
           </li>
           <li class="nav-item dropdown">
           <a href="#" class="avatar" id="navbarDropdown" data-bs-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
-            <%# <%= image_tag "https://kitt.lewagon.com/placeholder/users/ssaunier", class: "avatar", alt: "dropdown menu" %> 
+            <%# <%= image_tag "https://kitt.lewagon.com/placeholder/users/ssaunier", class: "avatar", alt: "dropdown menu" %>
           </a>
             <div class="dropdown-menu dropdown-menu-end" aria-labelledby="navbarDropdown">
               <%= link_to "Action", "#", class: "dropdown-item" %>
