+++ note
@@ -1,0 +1,68 @@
+import { Controller } from "@hotwired/stimulus"
+import Microphone from "../controllers/microphone_controller";
+export default class extends Controller {
+  connect() {
+    console.log("Connected.");
+
+    const main = () => {
+      const canvas = document.getElementById('myCanvas');
+      console.log(canvas);
+      const ctx = canvas.getContext('2d');
+      canvas.width = window.innerWidth;
+      canvas.height = window.innerHeight;
+
+      class Bar {
+        constructor(x, y, width, height, color) {
+          this.x = x;
+          this.y = y;
+          this.width = width;
+          this.height = height;
+          this.color = color;
+        }
+        update(micInput) {
+          this.height = micInput
+
+        }
+
+        draw(context) {
+          context.fillStyle = this.color;
+          context.fillRect(this.x, this.y, this.width, this.height);
+        }
+      }
+
+      // const bar1 = new Bar(10, 10, 100, 200, "blue");
+      const microphone = new Microphone();
+      let bars = [];
+      let barWidth = canvas.width/256;
+      function createBars() {
+          for (let i = 0; i < 256; i++){
+            let color = 'hsl(' + i * 2 + ' , 100%, 50%)';
+              bars.push( new Bar(i * barWidth, canvas.height/2, 1, 20, color));
+
+          }
+      }
+      createBars();
+      console.log(bars)
+
+      function animate() {
+
+        if (microphone.initialized){
+
+          ctx.clearRect(0, 0, canvas.width, canvas.height);
+          // get the audio samples
+          const samples = microphone.getSamples();
+          console.log(samples);
+          bars.forEach(function(bar, i) {
+            bar.update(i); // to make the pixxol heigt
+            bar.draw(ctx);
+          });
+
+        }
+        requestAnimationFrame(animate);
+      }
+
+      animate();
+    };
+    main();
+  }
+}
